chore: double click action added and code refactor
